Task-55442: Rename package name from org.exoplatform to io.meeds
