+++ poll-services/src/main/java/io/meeds/poll/entity/PollEntity.java
@@ -1,0 +1,113 @@
+/*
+ * This file is part of the Meeds project (https://meeds.io/).
+ * 
+ * Copyright (C) 2022 Meeds Association contact@meeds.io
+ * 
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package io.meeds.poll.entity;
+
+import org.exoplatform.commons.api.persistence.ExoEntity;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.Date;
+
+@Entity(name = "Poll")
+@ExoEntity
+@Table(name = "POLL")
+public class PollEntity implements Serializable {
+
+  private static final long serialVersionUID = 5290107403575974438L;
+
+  @Id
+  @SequenceGenerator(name = "SEQ_POLL_ID", sequenceName = "SEQ_POLL_ID", allocationSize = 1)
+  @GeneratedValue(strategy = GenerationType.AUTO, generator = "SEQ_POLL_ID")
+  @Column(name = "POLL_ID", nullable = false)
+  private Long              id;
+
+  @Column(name = "QUESTION")
+  private String            question;
+
+  @Column(name = "CREATED_DATE", nullable = false)
+  private Date              createdDate;
+
+  @Column(name = "END_DATE", nullable = false)
+  private Date              endDate;
+
+  @Column(name = "CREATOR_ID", nullable = false)
+  private Long              creatorId;
+
+  @Column(name = "ACTIVITY_ID")
+  private Long              activityId;
+  
+  @Column(name = "SPACE_ID")
+  private Long              spaceId;
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public String getQuestion() {
+    return question;
+  }
+
+  public void setQuestion(String question) {
+    this.question = question;
+  }
+
+  public Date getCreatedDate() {
+    return createdDate;
+  }
+
+  public void setCreatedDate(Date createdDate) {
+    this.createdDate = createdDate;
+  }
+
+  public Date getEndDate() {
+    return endDate;
+  }
+
+  public void setEndDate(Date endDate) {
+    this.endDate = endDate;
+  }
+
+  public Long getCreatorId() {
+    return creatorId;
+  }
+
+  public void setCreatorId(Long creatorId) {
+    this.creatorId = creatorId;
+  }
+
+  public Long getActivityId() {
+    return activityId;
+  }
+
+  public void setActivityId(Long activityId) {
+    this.activityId = activityId;
+  }
+
+  public Long getSpaceId() {
+    return spaceId;
+  }
+
+  public void setSpaceId(Long spaceId) {
+    this.spaceId = spaceId;
+  }
+}
